requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.0.0
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
xgboost>=2.0.0
category_encoders>=2.6.0
requests>=2.31.0
yfinance>=0.2.0
mysql-connector-python>=8.3.0
